remove duplicate code helpers from workflow-ui & builder
