+++ packages/frinx-workflow-builder/src/components/expanded-workflow-modal/expanded-workflow-modal.tsx
@@ -16,12 +16,11 @@ import {
   ModalHeader,
   ModalOverlay,
 } from '@chakra-ui/react';
-import { callbackUtils, Workflow, Task } from '@frinx/shared/src';
+import { callbackUtils, Workflow, Task, getElementsFromWorkflow, convertTaskToExtendedTask } from '@frinx/shared/src';
 import FeatherIcon from 'feather-icons-react';
 import React, { FC, useEffect, useMemo, useRef, useState } from 'react';
 import ReactFlow, { Background, BackgroundVariant, Controls, MiniMap, ReactFlowProvider } from 'react-flow-renderer';
 import { Link } from 'react-router-dom';
-import { convertTaskToExtendedTask, getElementsFromWorkflow } from '../../helpers/api-to-graph.helpers';
 import { getLayoutedElements } from '../../helpers/layout.helpers';
 import BaseNode from '../workflow-nodes/base-node';
 import DecisionNode from '../workflow-nodes/decision-node';
