+++ packages/shared/package.json
@@ -16,7 +16,8 @@
     "build": "run-s build:webpack generate:typings",
     "typegen": "tsc --declaration --jsx react --esModuleInterop true --target ESNext --moduleResolution node --outDir dist src/index.ts",
     "lint": "eslint --ext .ts,.tsx 'src/**/*.ts' 'src/**/*.tsx' --max-warnings 0",
-    "type-check": "tsc --noEmit"
+    "type-check": "tsc --noEmit",
+    "test": "vitest"
   },
   "devDependencies": {
     "@chakra-ui/react": "1.8.9",
@@ -40,8 +41,10 @@
     "npm-run-all": "4.1.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-flow-renderer": "10.3.17",
     "react-router-dom": "6.4.4",
-    "typescript": "4.8.4"
+    "typescript": "4.8.4",
+    "vitest": "0.25.3"
   },
   "dependencies": {
     "chakra-ui-autocomplete": "^1.4.5",
@@ -60,6 +63,7 @@
     "framer-motion": "^4.1.3 || ^6.0.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-flow-renderer": "10.3.17",
     "react-router-dom": "^6.0.0"
   }
 }
