+++ packages/shared/src/helpers/api-to-graph.helpers.ts
@@ -1,5 +1,6 @@
 import { Edge, Node } from 'react-flow-renderer';
 import { v4 as uuid } from 'uuid';
+import { getTaskLabel } from './task.helpers';
 import {
   DecisionTask,
   ExtendedDecisionTask,
@@ -13,8 +14,7 @@ import {
   NodeData,
   Task,
   TaskType,
-} from '@frinx/shared/src';
-import { getTaskLabel } from './task.helpers';
+} from './workflow-api.types';
 
 type NodeType = 'decision' | 'fork_join' | 'join' | 'base';
